# apidocs.yaml
openapi: 3.0.3
info:
  title: HISSA Market Analytics API
  version: 1.0.0
  description: |
    HISSA API v1 documentation based on the Laravel backend implementation.
    All endpoints are read-only, return JSON, and are available under the /api/v1 prefix.
    Dates use the YYYY-MM-DD format. Transaction values use IDR, while ratios use percent.
    The throttle:ip middleware limits requests per IP address according to the server configuration.
  license: { name: Proprietary }
servers:
  - url: /api/v1
    description: Active application server
...
paths:
...
  /market/dominance-ratio:
    get:
      tags: [Market Analysis]
      operationId: getDominanceRatio
      summary: Get investor dominance ratio
      description: Retrieves timeline institutional, retail, and mixed ratios for one stock. total_ratio is the sum of all three values.
      parameters:
        - $ref: "#/components/parameters/StartDate"
        - $ref: "#/components/parameters/EndDate"
        - name: stock_code
          in: query
          required: true
          description: Single stock code with a maximum length of 20 characters.
          schema: { type: string, maxLength: 20 }
        - name: granularity
          in: query
          description: Timeline resolution. Weekly and monthly return latest observation in each bucket.
          schema: { type: string, enum: [daily, weekly, monthly], default: daily }
      responses:
        "200":
          description: The dominance ratio was retrieved successfully.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DominanceRatioResponse" }
              example:
                stock_code: BBCA
                period: { start_date: "2026-07-01", end_date: "2026-07-01" }
                granularity: daily
                points:
                  [
                    {
                      date: "2026-07-01",
                      institution_ratio: 52.5,
                      retail_ratio: 35,
                      mixed_ratio: 12.5,
                      total_ratio: 100,
                    },
                  ]
                meta: { ratio_basis: transaction_value, unit: percent, aggregation: daily, timezone: Asia/Jakarta }
        "422": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalServerError" }
components:
  parameters:
    StartDate:
      name: start_date
      in: query
      required: true
      description: Start date of the period, inclusive.
      schema: { type: string, format: date }
      example: "2026-07-01"
    EndDate:
      name: end_date
      in: query
      required: true
      description: End date of the period, inclusive, and must not be earlier than start_date.
      schema: { type: string, format: date }
      example: "2026-07-22"
...
  responses:
    ValidationError:
      description: A path or query parameter failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationError" }
          example:
            success: false
            message: Validation failed.
            errors:
              end_date:
                [
                  The end date field must be a date after or equal to the start date.,
                ]
    TooManyRequests:
      description: The request exceeded the IP-based rate limit.
      headers:
        X-RateLimit-Limit:
          {
            description: Maximum requests in the window.,
            schema: { type: integer },
          }
        Retry-After:
          {
            description: Seconds recommended before retrying.,
            schema: { type: integer },
          }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ProblemMessage" }
          example: { message: Too Many Attempts. }
    InternalServerError:
      description: A query, domain process, or server operation failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ServerError" }
          example: { success: false, message: An unexpected error occurred. }
  schemas:
...
    DominanceRatioPoint:
      type: object
      required: [date, institution_ratio, retail_ratio, mixed_ratio, total_ratio]
      properties:
        date: { type: string, format: date }
        institution_ratio: { type: number, format: double, minimum: 0 }
        retail_ratio: { type: number, format: double, minimum: 0 }
        mixed_ratio: { type: number, format: double, minimum: 0 }
        total_ratio: { type: number, format: double }
    DominanceRatioResponse:
      type: object
      required: [stock_code, period, granularity, points, meta]
      properties:
        stock_code: { type: string }
        period:
          type: object
          required: [start_date, end_date]
          properties:
            start_date: { type: string, format: date }
            end_date: { type: string, format: date }
        granularity: { type: string, enum: [daily, weekly, monthly] }
        points:
          {
            type: array,
            items: { $ref: "#/components/schemas/DominanceRatioPoint" },
          }
        meta:
          type: object
          required: [ratio_basis, unit, aggregation, timezone]
          properties:
            ratio_basis: { type: string, enum: [transaction_value] }
            unit: { type: string, enum: [percent] }
            aggregation: { type: string, enum: [daily, latest] }
            timezone: { type: string, example: Asia/Jakarta }
    ValidationError:
      type: object
      required: [success, message, errors]
      properties:
        success: { type: boolean, enum: [false] }
        message: { type: string, enum: [Validation failed.] }
        errors:
          type: object
          additionalProperties:
            { type: array, minItems: 1, items: { type: string } }
    ServerError:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean, enum: [false] }
        message: { type: string }
    ProblemMessage:
      type: object
      required: [message]
      properties: { message: { type: string } }